New featers in the conflict category have been added

+++ apps/web/shared/lib/schema/conflict.ts
@@ -5,8 +5,8 @@ export const ConflictSchema = z.object({
     id: z.string().optional(),
     name: z.string(),
     description: z.string(),
-    startYear: z.number().optional(),
-    endYear: z.number().optional(),
+    startYear: z.number().int().optional(),
+    endYear: z.number().int().optional(),
     createdAt: z.number().optional(),
     updatedAt: z.number().optional(),
 })
